TIP-860: Move ValueNormalizer for Versioning from Tool to Pim/Enrichment
