+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/serializers_versionning.yml
@@ -1,6 +1,7 @@
 parameters:
     pim_serializer.normalizer.flat.category.class: Akeneo\Pim\Enrichment\Component\Category\Normalizer\Versioning\CategoryNormalizer
     pim_serializer.normalizer.flat.product.class: Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\ProductNormalizer
+    pim_serializer.normalizer.flat.value.class: Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product\ValueNormalizer
 
 services:
     pim_versioning.serializer.normalizer.flat.category:
@@ -13,5 +14,11 @@ services:
 
     pim_versioning.serializer.normalizer.flat.product:
         class: '%pim_serializer.normalizer.flat.product.class%'
+        tags:
+            - { name: pim_versioning.serializer.normalizer, priority: 90 }
+
+    pim_versioning.serializer.normalizer.flat.value:
+        public: false
+        class: '%pim_serializer.normalizer.flat.value.class%'
         tags:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
