+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Normalizer/Versioning/Product/ValueNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product;
 
 use Akeneo\Tool\Component\Localization\Localizer\NumberLocalizer;
 use Doctrine\Common\Collections\ArrayCollection;
@@ -10,6 +10,8 @@ use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeOptionInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ValueInterface;
 use Prophecy\Argument;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+use Symfony\Component\Serializer\SerializerAwareInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 
 class ValueNormalizerSpec extends ObjectBehavior
@@ -20,7 +22,7 @@ class ValueNormalizerSpec extends ObjectBehavior
     ) {
         $this->beConstructedWith(4);
 
-        $serializer->implement('Symfony\Component\Serializer\Normalizer\NormalizerInterface');
+        $serializer->implement(NormalizerInterface::class);
         $this->setSerializer($serializer);
 
         $simpleAttribute->isLocalizable()->willReturn(false);
@@ -30,8 +32,8 @@ class ValueNormalizerSpec extends ObjectBehavior
 
     function it_is_a_serializer_aware_normalizer()
     {
-        $this->shouldBeAnInstanceOf('Symfony\Component\Serializer\Normalizer\NormalizerInterface');
-        $this->shouldBeAnInstanceOf('Symfony\Component\Serializer\SerializerAwareInterface');
+        $this->shouldBeAnInstanceOf(NormalizerInterface::class);
+        $this->shouldBeAnInstanceOf(SerializerAwareInterface::class);
     }
 
     function it_supports_csv_normalization_of_product_value(ValueInterface $value)
